+++ src/libAtomVM/ets_hashtable.c
@@ -266,13 +266,28 @@ static uint32_t hash_external_pid(term t, int32_t h, GlobalContext *global)
     return h * LARGE_PRIME_PID;
 }
 
-static uint32_t hash_reference(term t, int32_t h, GlobalContext *global)
+static uint32_t hash_local_reference(term t, int32_t h, GlobalContext *global)
 {
     UNUSED(global);
     uint64_t n = term_to_ref_ticks(t);
     while (n) {
         h = h * LARGE_PRIME_REF + (n & 0xFF);
         n >>= 8;
+    }
+    return h * LARGE_PRIME_REF;
+}
+
+static uint32_t hash_external_reference(term t, int32_t h, GlobalContext *global)
+{
+    UNUSED(global);
+    uint32_t l = term_get_external_reference_len(t);
+    const uint32_t *words = term_get_external_reference_words(t);
+    for (uint32_t i = 0; i < l; i++) {
+        uint32_t n = words[i];
+        while (n) {
+            h = h * LARGE_PRIME_REF + (n & 0xFF);
+            n >>= 8;
+        }
     }
     return h * LARGE_PRIME_REF;
 }
@@ -300,8 +315,10 @@ static uint32_t hash_term_incr(term t, int32_t h, GlobalContext *global)
         return hash_local_pid(t, h, global);
     } else if (term_is_external_pid(t)) {
         return hash_external_pid(t, h, global);
-    } else if (term_is_reference(t)) {
-        return hash_reference(t, h, global);
+    } else if (term_is_local_reference(t)) {
+        return hash_local_reference(t, h, global);
+    } else if (term_is_external_reference(t)) {
+        return hash_external_reference(t, h, global);
     } else if (term_is_binary(t)) {
         return hash_binary(t, h, global);
     } else if (term_is_tuple(t)) {