+++ src/libAtomVM/term.c
@@ -23,6 +23,7 @@
 #include "atom.h"
 #include "atom_table.h"
 #include "context.h"
+#include "defaultatoms.h"
 #include "interop.h"
 #include "module.h"
 #include "tempstack.h"
@@ -31,6 +32,7 @@
 #include <inttypes.h>
 #include <stdarg.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 
 struct FprintfFun
@@ -384,11 +386,24 @@ int term_funprint(PrinterFun *fun, term t, const GlobalContext *global)
         ret += printed;
         return ret;
 
-    } else if (term_is_reference(t)) {
+    } else if (term_is_local_reference(t)) {
         uint64_t ref_ticks = term_to_ref_ticks(t);
 
         // Update also REF_AS_CSTRING_LEN when changing this format string
-        return fun->print(fun, "#Ref<0.0.0.%" PRIu64 ">", ref_ticks);
+        return fun->print(fun, "#Ref<0.%" PRIu32 ".%" PRIu32 ">", (uint32_t) (ref_ticks >> 32), (uint32_t) ref_ticks);
+
+    } else if (term_is_external_reference(t)) {
+        // Update also REF_AS_CSTRING_LEN when changing this format string
+        uint32_t node_atom_index = term_to_atom_index(term_get_external_node(t));
+        uint32_t len = term_get_external_reference_len(t);
+        const uint32_t *data = term_get_external_reference_words(t);
+        // creation is not printed
+        int ret = fun->print(fun, "#Ref<%" PRIu32, node_atom_index);
+        for (int i = len - 1; i >= 0; i--) {
+            ret += fun->print(fun, ".%" PRIu32, data[i]);
+        }
+        ret += fun->print(fun, ">");
+        return ret;
 
     } else if (term_is_boxed_integer(t)) {
         int size = term_boxed_size(t);
@@ -499,13 +514,67 @@ TermCompareResult term_compare(term t, term other, TermCompareOpts opts, GlobalC
             break;
 
         } else if (term_is_reference(t) && term_is_reference(other)) {
-            int64_t t_ticks = term_to_ref_ticks(t);
-            int64_t other_ticks = term_to_ref_ticks(other);
-            if (t_ticks == other_ticks) {
-                CMP_POP_AND_CONTINUE();
+            if (!term_is_external(t) && !term_is_external(other)) {
+                int64_t t_ticks = term_to_ref_ticks(t);
+                int64_t other_ticks = term_to_ref_ticks(other);
+                if (t_ticks == other_ticks) {
+                    CMP_POP_AND_CONTINUE();
+                } else {
+                    result = (t_ticks > other_ticks) ? TermGreaterThan : TermLessThan;
+                    break;
+                }
             } else {
-                result = (t_ticks > other_ticks) ? TermGreaterThan : TermLessThan;
-                break;
+                term node = term_is_external(t) ? term_get_external_node(t) : NONODE_AT_NOHOST_ATOM;
+                term other_node = term_is_external(other) ? term_get_external_node(other) : NONODE_AT_NOHOST_ATOM;
+                if (node == other_node) {
+                    uint32_t creation = term_is_external(t) ? term_get_external_node_creation(t) : 0;
+                    uint32_t other_creation = term_is_external(other) ? term_get_external_node_creation(other) : 0;
+                    if (creation == other_creation) {
+                        uint32_t len = term_is_external(t) ? term_get_external_reference_len(t) : 2;
+                        uint32_t other_len = term_is_external(other) ? term_get_external_reference_len(other) : 2;
+                        if (len == other_len) {
+                            const uint32_t *data;
+                            const uint32_t *other_data;
+                            uint32_t local_data[2];
+                            if (term_is_external(t)) {
+                                data = term_get_external_reference_words(t);
+                            } else {
+                                uint64_t ref_ticks = term_to_ref_ticks(t);
+                                local_data[0] = ref_ticks >> 32;
+                                local_data[1] = (uint32_t) ref_ticks;
+                                data = local_data;
+                            }
+                            if (term_is_external(other)) {
+                                other_data = term_get_external_reference_words(other);
+                            } else {
+                                uint64_t ref_ticks = term_to_ref_ticks(other);
+                                local_data[0] = ref_ticks >> 32;
+                                local_data[1] = (uint32_t) ref_ticks;
+                                other_data = local_data;
+                            }
+                            // Comparison is done in reverse order
+                            for (int i = len - 1; i >= 0; i--) {
+                                if (data[i] != other_data[i]) {
+                                    result = (data[i] > other_data[i]) ? TermGreaterThan : TermLessThan;
+                                    break;
+                                }
+                            }
+                            if (result != TermEquals) {
+                                break;
+                            }
+                            CMP_POP_AND_CONTINUE();
+                        } else {
+                            result = (len > other_len) ? TermGreaterThan : TermLessThan;
+                            break;
+                        }
+                    } else {
+                        result = (creation > other_creation) ? TermGreaterThan : TermLessThan;
+                        break;
+                    }
+                } else {
+                    t = node;
+                    other = other_node;
+                }
             }
 
         } else if (term_is_nonempty_list(t) && term_is_nonempty_list(other)) {
@@ -671,45 +740,36 @@ TermCompareResult term_compare(term t, term other, TermCompareOpts opts, GlobalC
             result = (atom_cmp_result > 0) ? TermGreaterThan : TermLessThan;
             break;
 
-        } else if (term_is_external_pid(t) && term_is_external_pid(other)) {
-            term node = term_get_external_node(t);
-            term other_node = term_get_external_node(other);
-            if (node == other_node) {
-                uint32_t creation = term_get_external_node_creation(t);
-                uint32_t other_creation = term_get_external_node_creation(other);
-                if (creation == other_creation) {
-                    uint32_t serial = term_get_external_pid_serial(t);
-                    uint32_t other_serial = term_get_external_pid_serial(other);
-                    if (serial == other_serial) {
-                        uint32_t process_id = term_get_external_pid_process_id(t);
-                        uint32_t other_process_id = term_get_external_pid_process_id(other);
-                        if (process_id == other_process_id) {
+        } else if (term_is_pid(t) && term_is_pid(other)) {
+            uint32_t process_id = term_is_external(t) ? term_get_external_pid_process_id(t) : (uint32_t) term_to_local_process_id(t);
+            uint32_t other_process_id = term_is_external(other) ? term_get_external_pid_process_id(other) : (uint32_t) term_to_local_process_id(other);
+            if (process_id == other_process_id) {
+                uint32_t serial = term_is_external(t) ? term_get_external_pid_serial(t) : 0;
+                uint32_t other_serial = term_is_external(other) ? term_get_external_pid_serial(other) : 0;
+                if (serial == other_serial) {
+                    term node = term_is_external(t) ? term_get_external_node(t) : NONODE_AT_NOHOST_ATOM;
+                    term other_node = term_is_external(other) ? term_get_external_node(other) : NONODE_AT_NOHOST_ATOM;
+                    if (node == other_node) {
+                        uint32_t creation = term_is_external(t) ? term_get_external_node_creation(t) : 0;
+                        uint32_t other_creation = term_is_external(other) ? term_get_external_node_creation(other) : 0;
+                        if (creation == other_creation) {
                             CMP_POP_AND_CONTINUE();
                         } else {
-                            result = (process_id > other_process_id) ? TermGreaterThan : TermLessThan;
+                            result = (creation > other_creation) ? TermGreaterThan : TermLessThan;
                             break;
                         }
                     } else {
-                        result = (serial > other_serial) ? TermGreaterThan : TermLessThan;
-                        break;
+                        t = node;
+                        other = other_node;
                     }
                 } else {
-                    result = (creation > other_creation) ? TermGreaterThan : TermLessThan;
+                    result = (serial > other_serial) ? TermGreaterThan : TermLessThan;
                     break;
                 }
             } else {
-                result = (node > other_node) ? TermGreaterThan : TermLessThan;
+                result = (process_id > other_process_id) ? TermGreaterThan : TermLessThan;
                 break;
             }
-        } else if (term_is_local_pid(t) && term_is_local_pid(other)) {
-            //TODO: handle ports
-            result = (t > other) ? TermGreaterThan : TermLessThan;
-            break;
-
-        } else if (term_is_pid(t) && term_is_pid(other)) {
-            result = term_is_local_pid(other) ? TermGreaterThan : TermLessThan;
-            break;
-
         } else {
             result = (term_type_to_index(t) > term_type_to_index(other)) ? TermGreaterThan : TermLessThan;
             break;