+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `erlang:split_binary/2`
 - Added `inet:getaddr/2`
 - Added support for external pids and encoded pids in external terms
+- Added support for external refs and encoded refs in external terms
 
 ## [0.6.6] - Unreleased